% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_varnames.R
\name{get_varnames}
\alias{get_varnames}
\title{Gets possible varnames for FLUXNET FULLSET/SUBSET and La Thuile}
\usage{
get_varnames(datasetname, flx2015_version = "FULLSET", add_psurf = FALSE)
}
\arguments{
\item{datasetname}{one of "FLUXNET2015", "LaThuile" and "OzFlux"}

\item{flx2015_version}{(optional) only for FLUXNET2015, one of "SUBSET" and "FULLSET"}

\item{add_psurf}{(optional) only for LaThuile}
}
\description{
Gets possible varnames for FLUXNET FULLSET/SUBSET and La Thuile
}
\examples{
get_varnames("FLUXNET2015")
}
