% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LaThuile_site_policy}
\alias{LaThuile_site_policy}
\title{DATASET_TITLE}
\format{
A data frame with 253 rows and 18 variables:
\describe{
  \item{`V1`}{character COLUMN_DESCRIPTION}
  \item{`V2`}{character COLUMN_DESCRIPTION}
  \item{`V3`}{character COLUMN_DESCRIPTION}
  \item{`V4`}{character COLUMN_DESCRIPTION}
  \item{`V5`}{character COLUMN_DESCRIPTION}
  \item{`V6`}{character COLUMN_DESCRIPTION}
  \item{`V7`}{character COLUMN_DESCRIPTION}
  \item{`V8`}{character COLUMN_DESCRIPTION}
  \item{`V9`}{character COLUMN_DESCRIPTION}
  \item{`V10`}{character COLUMN_DESCRIPTION}
  \item{`V11`}{character COLUMN_DESCRIPTION}
  \item{`V12`}{character COLUMN_DESCRIPTION}
  \item{`V13`}{character COLUMN_DESCRIPTION}
  \item{`V14`}{character COLUMN_DESCRIPTION}
  \item{`V15`}{character COLUMN_DESCRIPTION}
  \item{`V16`}{character COLUMN_DESCRIPTION}
  \item{`V17`}{character COLUMN_DESCRIPTION}
  \item{`V18`}{character COLUMN_DESCRIPTION} 
}
}
\source{
<http://somewhere.important.com/>
}
\usage{
LaThuile_site_policy
}
\description{
DATASET_DESCRIPTION
}
\examples{
\dontrun{
library(ncdf4)
library(magrittr)

file   = "OUTPUTS/Nc_files/Flux/AU-Whr_2012-2014_FLUXNET2015_Flux.nc"
ncfile = nc_open(file)

vars_ndim <- map_int(ncfile$var, ~length(.x$varsize))
vars <- vars_ndim \%>\% {names(.)[. == 3]} \%>\% print()

plot_nc(ncfile, analysis_type = c("diurnal"), vars, outfile = "diurnal_")
# SumatraPDF("diurnal_DiurnalCycle.pdf")
}
}
\keyword{datasets}
