% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Output_variables_FLUXNET2015_FULLSET}
\alias{Output_variables_FLUXNET2015_FULLSET}
\title{Output_variables_FLUXNET2015_FULLSET}
\format{A data frame with 40 rows and 15 variables:
\describe{
  \item{`Fluxnet_variable`}{character COLUMN_DESCRIPTION}
  \item{`Fluxnet_unit`}{character COLUMN_DESCRIPTION}
  \item{`Fluxnet_class`}{character COLUMN_DESCRIPTION}
  \item{`Output_variable`}{character COLUMN_DESCRIPTION}
  \item{`Output_unit`}{character COLUMN_DESCRIPTION}
  \item{`Longname`}{character COLUMN_DESCRIPTION}
  \item{`Standard_name`}{character COLUMN_DESCRIPTION}
  \item{`CMIP_short_name`}{character COLUMN_DESCRIPTION}
  \item{`Data_min`}{integer COLUMN_DESCRIPTION}
  \item{`Data_max`}{double COLUMN_DESCRIPTION}
  \item{`Essential_met`}{integer COLUMN_DESCRIPTION}
  \item{`Preferred_eval`}{logical COLUMN_DESCRIPTION}
  \item{`Category`}{character COLUMN_DESCRIPTION}
  \item{`ERAinterim_variable`}{character COLUMN_DESCRIPTION}
  \item{`Aggregate_method`}{character COLUMN_DESCRIPTION} 
}}
\source{
<http://somewhere.important.com/>
}
\usage{
Output_variables_FLUXNET2015_FULLSET
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
