% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfluxnet-package.R
\docType{package}
\name{rfluxnet-package}
\alias{rfluxnet}
\alias{rfluxnet-package}
\title{rfluxnet: Process fluxnet data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com}

}
\keyword{internal}
