% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{site_metadata}
\alias{site_metadata}
\title{DATASET_TITLE}
\format{
A data frame with 856 rows and 22 variables:
\describe{
  \item{`SiteCode`}{character COLUMN_DESCRIPTION}
  \item{`Exclude`}{character COLUMN_DESCRIPTION}
  \item{`Exclude_reason`}{logical COLUMN_DESCRIPTION}
  \item{`CABLE_PFT`}{integer COLUMN_DESCRIPTION}
  \item{`Description`}{character COLUMN_DESCRIPTION}
  \item{`TowerStatus`}{character COLUMN_DESCRIPTION}
  \item{`Country`}{character COLUMN_DESCRIPTION}
  \item{`Fullname`}{character COLUMN_DESCRIPTION}
  \item{`SiteLatitude`}{double COLUMN_DESCRIPTION}
  \item{`SiteLongitude`}{double COLUMN_DESCRIPTION}
  \item{`SiteElevation`}{double COLUMN_DESCRIPTION}
  \item{`IGBP_vegetation_short`}{character COLUMN_DESCRIPTION}
  \item{`IGBP_vegetation_long`}{character COLUMN_DESCRIPTION}
  \item{`Tier`}{integer COLUMN_DESCRIPTION}
  \item{`MeasurementHeight`}{double COLUMN_DESCRIPTION}
  \item{`TowerHeight`}{character COLUMN_DESCRIPTION}
  \item{`CanopyHeight`}{double COLUMN_DESCRIPTION}
  \item{`VegetationDescription`}{character COLUMN_DESCRIPTION}
  \item{`SoilType`}{character COLUMN_DESCRIPTION}
  \item{`Disturbance`}{character COLUMN_DESCRIPTION}
  \item{`CropDescription`}{character COLUMN_DESCRIPTION}
  \item{`Irrigation`}{character COLUMN_DESCRIPTION}
}
}
\source{
<http://somewhere.important.com/>
}
\usage{
site_metadata
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
